requirements.txt
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.51
botocore==1.17.51
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
gensim==3.8.3
idna==2.10
IMDbPY==6.8
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.1
python-dateutil==2.8.1
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0
smart-open==2.1.1
soupsieve==2.0.1
SQLAlchemy==1.3.19
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
